began debugging initialization; fixed init code; still a bug in probe
shell scripts were updated to reflect new directory structure
the code in the init funciton has been tested and works.  The code in probe
dies because I did not account for the difference in xenstore between
the backend and frontend path.  Either xenstore needs to be initialized
to have the mount information in both the back and front end xenstore paths,
(this is preferable), or the accessing of the mount path has to be delayed
until after the back and front end are connected.

Signed-off-by: Linda Jacobson